ai_server/requirements.txt
flask==3.1.0
python-dotenv==1.0.1
torch==2.8.0
transformers==4.56.1
pillow==11.3.0
requests==2.32.3
qwen-vl-utils==0.0.11